Fix compile errors when <vulkan/vulkan.hpp> was included.
I really don't like it that you have to include <vulkan/vulkan.hpp>
BEFORE including a glfwpp header in order to get this C++ API;
nevertheless, this was clearly never tested.

I opted to duplicate the function, once with and once without
allocator. Seems more logical than to use a pointer that is nullptr
to signal you don't want an allocator (in C++).